Change ImageParams to use size_t for image dimensions and adjust everything that comes out of it

+++ yrt-pet/include/yrt-pet/datastruct/image/Image.hpp
@@ -28,7 +28,7 @@ public:
 	float* getRawPointer();
 	const float* getRawPointer() const;
 	bool isMemoryValid() const;
-	int getNumFrames() const;
+	size_t getNumFrames() const;
 
 	void copyFromImage(const ImageBase* imSrc) override;
 	void multWithScalar(float scalar);
